+++ hal/src/gcc/eeprom_hal.cpp
@@ -19,7 +19,7 @@
 
 #include "eeprom_hal.h"
 #include "eeprom_file.h"
-#include "filesystem.h"
+#include "filesystem_util.h"
 #include <string.h>
 #include <string>
