+++ hal/src/gcc/filesystem_util.h
